Major legacy cleanup and improvements (#951)
Improved GCD calculations
* Previous implementation essentially floored the gcd subtraction.
* New way rounds it and returns rotations closer to the original.

Converted RotationUtils to kotlin
* added GCD getter - getFixedAngleDelta
* added angle to GCD convertor - getFixedSensitivityAngle(targetAngle, startAngle = 0f)
* simplified EntityFilter interface to a lambda function

Added new player extensions
* mc.thePlayer.setFixedSensitivityAngles(yaw: Float?, pitch: Float?)
* mc.thePlayer.fixedSensitivityYaw
* mc.thePlayer.fixedSensitivityPitch
-> Changing these will set rotationYaw/rotationPitch to GCD patched angles.
-> You can even do `fixedSensitivityYaw += x`.

Improved code of modules
* implemented MovementUtils
* fixed GCD with use of new utils (Aimbot, AutoClicker, AntiAFK, AutoPot, Derp)
* simplified stuff

Rewrote many lines
* simplified null checks, when codeblocks
* replaced deprecated/java functions
* removed unnecessary `!!`, `?`, `;`, `throws`
* fixed typos
* supervised intellij suggestions, ...

+++ src/main/java/net/ccbluex/liquidbounce/utils/render/RenderUtils.java
@@ -496,9 +496,9 @@ public final class RenderUtils extends MinecraftInstance {
         Tessellator tessellator = Tessellator.getInstance();
         WorldRenderer worldrenderer = tessellator.getWorldRenderer();
         worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
-        worldrenderer.pos(x, y + height, 0.0D).tex(u * f, (v + (float) height) * f1).endVertex();
-        worldrenderer.pos(x + width, y + height, 0.0D).tex((u + (float) width) * f, (v + (float) height) * f1).endVertex();
-        worldrenderer.pos(x + width, y, 0.0D).tex((u + (float) width) * f, v * f1).endVertex();
+        worldrenderer.pos(x, y + height, 0.0D).tex(u * f, (v + height) * f1).endVertex();
+        worldrenderer.pos(x + width, y + height, 0.0D).tex((u + width) * f, (v + height) * f1).endVertex();
+        worldrenderer.pos(x + width, y, 0.0D).tex((u + width) * f, v * f1).endVertex();
         worldrenderer.pos(x, y, 0.0D).tex(u * f, v * f1).endVertex();
         tessellator.draw();
     }
